requirements.txt
cryptography==46.0.3
deepface==0.0.95
mtcnn==1.0.0
mysql_connector_repackaged==0.3.1
numpy==2.3.5
opencv_python==4.12.0.88
Pillow==12.0.0
pyserial==3.5
python-dotenv==1.2.1
ttkbootstrap==1.18.1
